Update space invite entry points
